Properly resolve nested imports in JSONSchema importer
Test were failing while importing the HCA project after updating
linkml-runtime to 1.9.x because the schema for the file core ontology
was trying to import this file:

tests/outputs/hca/core/file/module/ontology/file_content_ontology.yaml

when it was trying to import this file:

tests/outputs/hca/module/ontology/file_content_ontology.yaml

This was happening because it was importing
"module/ontology/file_content_ontology.yaml" and expected to be
importing from the root of the project. This was the behavior of linkml
runtime before 1.9.x, but that changed in this PR, which was merged and
released with 1.9.1: <https://github.com/linkml/linkml-runtime/pull/368>

The fix is to make this a relative path-- that is, the value
`../../module/ontology/file_content_ontology.yaml`. Luckily, the value
of the relative prefix was already in the code! It just wasn't being
appended to the module path. This might have been an oversight that
wasn't added since the code worked.